"UserId from Integer to Long"

+++ simpleGETproject/simpleget/src/main/java/com/alexdev/simpleget/controllers/UserController.java
@@ -17,7 +17,7 @@ public class UserController {
         return userService.getUsers();
     }
     @GetMapping("/user/{userId}")
-    public User getUserById(@PathVariable Integer userId){
+    public User getUserById(@PathVariable Long userId){
         return userService.getUserById(userId);
     }
 
@@ -26,11 +26,11 @@ public class UserController {
         userService.saveUser(newUser);
     }
     @DeleteMapping("/user/{userId}")
-    public void deleteUserById(@PathVariable Integer userId){
+    public void deleteUserById(@PathVariable Long userId){
         userService.deleteUserById(userId);
     }
     @PutMapping("/user/{userId}")
-    public void updateUser(@RequestBody User updateUser, @PathVariable Integer userId){
+    public void updateUser(@RequestBody User updateUser, @PathVariable Long userId){
         userService.saveUser(User.builder()
                 .userId(userId)
                 .dni(updateUser.getDni())
